Rename tabs_open to open_tabs
