+++ lib/sanbase_web/channels/user_channel.ex
@@ -2,14 +2,21 @@ defmodule SanbaseWeb.UserChannel do
   use SanbaseWeb, :channel
 
   alias SanbaseWeb.Presence
+  alias Sanbase.Accounts.User
 
-  def join("users:" <> _user_id, _params, socket) do
-    {:ok, _} = Presence.track(socket, socket.assigns.user_id, %{})
-    {:ok, socket}
+  def join("users:" <> user_id, _params, socket) do
+    case String.to_integer(user_id) == socket.assigns.user_id do
+      true ->
+        {:ok, _} = Presence.track(socket, socket.assigns.user_id, %{})
+        {:ok, socket}
+
+      false ->
+        {:error, "Cannot join other users rooms"}
+    end
   end
 
   def handle_in("is_username_valid", %{"username" => username}, socket) do
-    case Sanbase.Accounts.User.Username.valid?(username) do
+    case User.Username.valid?(username) do
       true ->
         {:reply, {:ok, %{"is_username_valid" => true}}, socket}
 
@@ -18,15 +25,15 @@ defmodule SanbaseWeb.UserChannel do
     end
   end
 
-  def handle_in("tabs_open", %{}, socket) do
+  def handle_in("open_tabs", %{}, socket) do
     user_id = "#{socket.assigns.user_id}"
 
     case Presence.list(socket) do
       %{^user_id => %{metas: list}} ->
-        {:reply, {:ok, %{"tabs_open" => length(list)}}, socket}
+        {:reply, {:ok, %{"open_tabs" => length(list)}}, socket}
 
       %{} = empty_map when map_size(empty_map) == 0 ->
-        {:reply, {:ok, %{"tabs_open" => 0}}, socket}
+        {:reply, {:ok, %{"open_tabs" => 0}}, socket}
     end
   end
 end
